Revert "Feat scanner"
