--- modules/multer.js
@@ -1,15 +1,0 @@
-const multer = require('multer')
-
-const storage = multer.diskStorage({
-    destination: (req, file, cb) => {
-        cb(null, './uploads')
-    },
-    filename: (req, file, cb) => {
-        cb(null, file.originalname)
-    }
-});
-const upload = multer({
-    storage: storage
-}).single('scanmeds')
-
-module.exports = upload
