+++ package.json
@@ -35,10 +35,6 @@
     "multer": "^1.4.2",
     "node-fetch": "^2.6.0",
     "nodemon": "^2.0.3",
-    "string-similarity": "^4.0.1",
     "tesseract.js": "^2.1.1"
-  },
-  "devDependencies": {
-    "string-similarity": "^4.0.1"
   }
 }
