--- routes/home.js
@@ -1,3 +1,0 @@
-module.exports = async (req, res) => {
-    res.render('index')
-}
