+++ views/pages/test.ejs
@@ -1,15 +1,2 @@
-<h2>Medicine</h2>
-<p> <%= text %> </p>
-
-<ul>
-    <li>
-        <%= meds.name %>
-    </li>
-    <li>
-        <%= meds.registrationNumber %>
-    </li>
-    <li>
-        <%= meds.activeIngredient %>
-    </li>
-
-</ul>
+<h2>testing</h2>
+<p>items </p>
