--- views/search.ejs
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Medicijnen scanner</title>
-</head>
-
-<body>
-    <form action="/search" method="POST">
-        <input type="text" name="medicine" id="medicine">
-        <button type="submit">search</button>
-    </form>
-</body>
-
-</html>
